+++ packages/react-components/src/components/RichTextEditor/Toolbar/RichTextToolbar.tsx
@@ -1,0 +1,20 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+import React from 'react';
+
+/**
+ * Props for {@link RichTextToolbar}.
+ *
+ * @private
+ */
+export interface RichTextToolbarProps {}
+
+/**
+ * A component to display rich text toolbar.
+ *
+ * @beta
+ */
+export const RichTextToolbar = (props: RichTextToolbarProps): JSX.Element => {
+  const {} = props;
+  return <div></div>;
+};
